Introduced a more readable class for embedding.
The old class had serious problems as it got bigger. This time, we try
to move as much logic to separate classes so we can take them out when
they become to big. Also, this time, we have loads of small methods that
essentially move logic over to helpers. This allows for smaller
functions and a better overview.